fix: Resolved all build warnings (#138)
* fix: Resolved all build warnings

* fix: Resolved current compiler warnings

* fix: Resolved another warning the compiler doesn't tell

* build: Added -Werror compiler flag

* build: Added -Wconversion compiler flag

* fix: Resolver all compiler warnings

+++ Drakhtar/GameObjects/Board.h
@@ -30,7 +30,7 @@ class Board final : public GameObject {
    * \brief The size of each cell of the board (Applies to both width and
    * height).
    */
-  float cellSize_;
+  double cellSize_;
 
   /**
    * \brief Auxiliary matrix that saves the last search of cells near another
@@ -39,7 +39,7 @@ class Board final : public GameObject {
   Matrix<ObjectType> *objectTypeMatrix_ = nullptr;
 
  public:
-  Board(Scene *scene, int rows, int columns, float cellSize);
+  Board(Scene *scene, int rows, int columns, double cellSize);
   virtual ~Board();
 
   /**
